style: :lipstick: 자잘한 디자인 관련 요소 수정
