+++ src/components/SecretPage/StatusSection/JobStats/index.tsx
@@ -1,8 +1,14 @@
 import { FrontIcon } from '@/components/SecretPage/assets';
 import { RangeIcon } from './RangeIcon';
-import { Keywords } from './Keywords';
 
-import { jobIconStyle, jobStatDivStyle, jobStatWrapperStyle, rangeWrapperStyle } from './JobStats.styles';
+import {
+  jobIconStyle,
+  jobStatDivStyle,
+  jobStatWrapperStyle,
+  rangeWrapperStyle,
+  keywordInner,
+  keywordWrapper,
+} from './JobStats.styles';
 
 export const JobStats = () => {
   return (
@@ -12,9 +18,13 @@ export const JobStats = () => {
       </div>
       <div css={[jobStatDivStyle, rangeWrapperStyle]}>
         <RangeIcon />
-        <span>공격 범위</span>
+        <span>커밋 범위</span>
       </div>
-      <Keywords />
+      <div css={keywordWrapper}>
+        <div css={[jobStatDivStyle, keywordInner]}>
+          <span>UI Design & Background Asset ⓒ Hypergryph</span>
+        </div>
+      </div>
     </div>
   );
 };
